Provide SHA512/256 crypto algorithm for Algorand Address encoding
