+++ ledger-core-algorand/inc/algorand/utils/SHA512256.hpp
@@ -33,16 +33,20 @@
 #include <vector>
 
 namespace ledger {
-    namespace core {
-        class SHA512256 {
-        public:
-            static std::string stringToHexHash(const std::string& input);
-            static std::string bytesToHexHash(const std::vector<uint8_t>& bytes);
-            static std::vector<uint8_t> stringToBytesHash(const std::string& input);
-            static std::vector<uint8_t> bytesToBytesHash(const std::vector<uint8_t>& bytes);
+namespace core {
+namespace algorand {
 
-        private:
-            static std::vector<uint8_t> dataToBytesHash(const void *data, size_t size);
-        };
-    }
+    class SHA512256 {
+    public:
+        static std::string stringToHexHash(const std::string& input);
+        static std::string bytesToHexHash(const std::vector<uint8_t>& bytes);
+        static std::vector<uint8_t> stringToBytesHash(const std::string& input);
+        static std::vector<uint8_t> bytesToBytesHash(const std::vector<uint8_t>& bytes);
+
+    private:
+        static std::vector<uint8_t> dataToBytesHash(const uint8_t *data, size_t size);
+    };
+
+}
+}
 }
